mqttdir/requirements.txt
pandas
paho-mqtt
redis